functions/requirements.txt
firebase-functions>=0.4.0
firebase-admin>=6.2.0
google-cloud-firestore>=2.13.0
pandas>=2.0.0
textblob>=0.18.0
nltk>=3.8.0

